chore: Type 패키지 정리
