+++ presentation/src/main/java/com/sopt/presentation/ui/screen/home/composable/HomeContentScreen.kt
@@ -14,6 +14,8 @@ import androidx.compose.ui.res.stringResource
 import androidx.compose.ui.unit.dp
 import com.sopt.presentation.R
 import com.sopt.presentation.ui.component.text.SecondaryText
+import com.sopt.presentation.ui.screen.home.type.VideoType
+import com.sopt.presentation.ui.state.VideoOverviewViewState
 import com.sopt.presentation.ui.theme.WavveTheme
 import com.sopt.presentation.util.noRippleClickable
 
@@ -52,18 +54,4 @@ fun VideoTypeTabRow(
             )
         }
     }
-}
-
-enum class VideoType(
-    @StringRes val titleResId: Int
-) {
-    NEW_CLASSIC(R.string.type_new_classic),
-    DRAMA(R.string.type_drama),
-    ENTERTAINMENT(R.string.type_entertainment),
-    MOVIE(R.string.type_movie),
-    ANIMATION(R.string.type_animation),
-    ABROAD_SERIES(R.string.type_abroad_series),
-    INFORMATION_CULTURE(R.string.type_information_culture),
-    KIDS(R.string.type_kids),
-    MOVIE_PLUS(R.string.type_movie_plus),
 }
